requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1